[ADD] Includes function tha creates metrics table
It can be exported to docx, pdf, png, and tex.
Ironically, docx are a better format than png and pdf, as it does not
requires to trimm the image to fits on the device (with webshot2).
The tex files seems to be good, but not compilation was tested so far.